Add files via upload
The Car Rental System is a web-based application designed to streamline the process of renting cars. This project provides a user-friendly interface for customers to browse available vehicles, make reservations, and manage their rentals. Built with HTML and CSS for the frontend, PHP for the backend, and MySQL for data management, this system aims to enhance the car rental experience for both customers and administrators.

Features